Updating Spring Boot

+++ backend-javaspring/src/main/java/org/example/app/models/responses/RenterRentCountDTO.java
@@ -1,3 +1,3 @@
 package org.example.app.models.responses;
 
-public record RenterRentCountDTO(String name, int rentsQuantity, int rentsActive) {}
+public record RenterRentCountDTO(String name, Long rentsQuantity, Long rentsActive) {}
